+++ apis/serve.go
@@ -156,13 +156,6 @@ func Serve(app core.App, config ServeConfig) error {
 		ErrorLog: log.New(&serverErrorLogWriter{app: app}, "", 0),
 	}
 
-	serveEvent := new(core.ServeEvent)
-	serveEvent.App = app
-	serveEvent.Router = pbRouter
-	serveEvent.Server = server
-	serveEvent.CertManager = certManager
-	serveEvent.InstallerFunc = DefaultInstallerFunc
-
 	var listener net.Listener
 
 	// graceful shutdown
@@ -210,6 +203,13 @@ func Serve(app core.App, config ServeConfig) error {
 
 	var baseURL string
 
+	serveEvent := new(core.ServeEvent)
+	serveEvent.App = app
+	serveEvent.Router = pbRouter
+	serveEvent.Server = server
+	serveEvent.CertManager = certManager
+	serveEvent.InstallerFunc = DefaultInstallerFunc
+
 	// trigger the OnServe hook and start the tcp listener
 	serveHookErr := app.OnServe().Trigger(serveEvent, func(e *core.ServeEvent) error {
 		// Support path prefix for the API and dashboard routes.
@@ -245,9 +245,13 @@ func Serve(app core.App, config ServeConfig) error {
 			}
 		}
 
-		listener, err = net.Listen("tcp", addr)
-		if err != nil {
-			return err
+		if e.Listener == nil {
+			listener, err = net.Listen("tcp", addr)
+			if err != nil {
+				return err
+			}
+		} else {
+			listener = e.Listener
 		}
 
 		if e.InstallerFunc != nil {
@@ -268,7 +272,7 @@ func Serve(app core.App, config ServeConfig) error {
 
 	if listener == nil {
 		//nolint:staticcheck
-		return errors.New("The OnServe finalizer wasn't invoked. Did you forget to call the ServeEvent.Next() method?")
+		return errors.New("The OnServe listener was not initialized. Did you forget to call the ServeEvent.Next() method?")
 	}
 
 	if config.ShowStartBanner {